requirements.txt
yagmail==0.5.148
docopt==0.6.2
python-dateutil==2.6.0
retrying==1.3.3